Created Calculation HTML Page, Controller, Links
IT'S BY NO MEANS FINISHED
I took the <%= javascript_include_tag "application",
"data-turbolinks-track" => true %> out of the calculationpage in
app->views->layouts to make the javascript(adding/removing inputs) work
for the input form fyi
Still needs a lot of validation, and I'm not sure if I have enough input
fields for what we're trying to do.
As of right now, it is supposed to take all the values and stores them
in a two-dimensional array via javascript.
Looks like there's a couple things wrong with it:

Still need to account for:
if the user deletes rows in the middle of all the rows(messing up the
for loop)
getting the for loop to display alerts for sums of each row after the
first row.